+++ templates/master-schedule-table.php
@@ -13,14 +13,15 @@ $today = radio_station_get_time( 'day', $now );
 // 2.3.3.9: added for non-now schedule displays
 if ( isset( $atts['start_date'] ) && $atts['start_date'] ) {
 	$start_date = $atts['start_date'];
-	$start_time = radio_station_to_time( $start_date . ' 00:00:00' );
 	// --- force display of date and month ---
 	$atts['display_date'] = ( !$atts['display_date'] ) ? '1' : $atts['display_date'];
 	$atts['display_month'] = ( !$atts['display_month'] ) ? 'short' : $atts['display_month'];
 } else {
-	$start_time = $now;
+	// 2.3.3.9: set start date to current date
+	$start_date = $date;	
 }
-$start_time = apply_filters( 'radio_station_schedule_start_time', $start_time, 'table' );
+$start_time = radio_station_to_time( $start_date . ' 00:00:00' );
+$start_time = apply_filters( 'radio_station_schedule_start_time', $start_time, 'table', $atts );
 
 // --- set shift time formats ---
 // 2.3.2: set time formats early
@@ -79,7 +80,11 @@ $table .= '<table id="master-program-schedule" cellspacing="0" cellpadding="0">'
 // --- weekday table headings row ---
 // 2.3.2: added hour column heading id
 $table .= '<tr class="master-program-day-row">' . $newline;
-$table .= '<th id="master-program-hour-heading"></th>' . $newline;
+$table .= '<th id="master-program-hour-heading">' . $newline;
+// 2.3.3.9: added filters for week loader controls
+$table .= apply_filters( 'radio_station_schedule_loader_control', '', 'table', 'left' );
+$table .= apply_filters( 'radio_station_schedule_loader_control', '', 'table', 'right' );
+$table .= '</th>' . $newline;
 
 foreach ( $weekdays as $i => $weekday ) {
 
@@ -699,7 +704,7 @@ foreach ( $hours as $hour ) {
 					if ( !isset( $cell ) ) {
 						$cell = '';
 					}
-					$cell .= $add_link;
+					$cell .= '<center>' . $add_link . '</center>';
 				}
 			}
 			$cellclass = implode( ' ', $cellclasses );